Final update

+++ Sophomore Year/Computing 4/ps5/TFractal.cpp
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <vector>
 
+#include <math.h>
 #include <SFML/System.hpp>
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
@@ -28,7 +29,7 @@ int main(int argc, char* argv[]) {
             if (event.type == sf::Event::Closed)
                 window.close();
         window.clear();
-        for (int i = 0; i < triangleVec.size(); i++)
+        for (long unsigned int i = 0; i < triangleVec.size(); i++)
             window.draw(triangleVec[i].triangle);
         window.display();
     }
